V2

+++ Firmware/src/Sensors/vrailmonitor.cpp
@@ -1,4 +1,4 @@
-#include "VRailMonitor.h"
+#include "vrailmonitor.h"
 
 #include <string>
 
@@ -10,20 +10,22 @@
 #include <libriccore/riccorelogging.h>
 
 
+
+
 VRailMonitor::VRailMonitor(std::string_view vrail_name,const uint8_t pin, const float r1,const float r2):
-_name(vrail_name),
-_pin(pin),
-_channel(ADC_CHANNEL_0),//default
-_unit(ADC_UNIT_1),
-_adcCal(),
-_adcInitialized(false),
-factor(((r1+r2)/r2)),
-_maxVoltage(0),
-_lowVoltage(0),
-_minVoltage(0)
+    _name(vrail_name),
+    _pin(pin),
+    _channel(ADC_CHANNEL_0),//default
+    _unit(ADC_UNIT_1),
+    _adcCal(),
+    _adcInitialized(false),
+    factor(((r1+r2)/r2)),
+    _maxVoltage(0),
+    _lowVoltage(0),
+    _minVoltage(0)
 {};
 
-void VRailMonitor::setup(uint16_t maxVoltage, uint16_t lowVoltage,uint16_t minVoltage){
+void VRailMonitor::setup(int maxVoltage, int lowVoltage,int minVoltage){
     _maxVoltage = maxVoltage;
     _lowVoltage = lowVoltage;
     _minVoltage = minVoltage;
@@ -59,7 +61,7 @@ void VRailMonitor::setup(uint16_t maxVoltage, uint16_t lowVoltage,uint16_t minVo
     
 }
 
-void VRailMonitor::update(float &data)
+void VRailMonitor::update(float &data)    
 {
     if(!_adcInitialized)
     {
@@ -92,7 +94,6 @@ void VRailMonitor::update(float &data)
         {
             _lowVoltageTriggered = false;
         }
-        
     }
 }
